Flytt funksjon for å fjerne barn til context hvor legg til/endre skjer
