+++ src/context/SøknadContext.tsx
@@ -200,6 +200,17 @@ const [SøknadProvider, useSøknad] = createUseContext(() => {
     }));
   };
 
+  const fjernBarnFraSøknad = (id: string) => {
+    const nyBarneListe = søknad.person.barn.filter((b: IBarn) => b.id !== id);
+
+    settSøknad((prevSoknad: ISøknad) => {
+      return {
+        ...prevSoknad,
+        person: { ...søknad.person, barn: nyBarneListe },
+      };
+    });
+  };
+
   return {
     søknad,
     settSøknad,
@@ -212,6 +223,7 @@ const [SøknadProvider, useSøknad] = createUseContext(() => {
     nullstillMellomlagretOvergangsstønad,
     oppdaterBarnISoknaden,
     nullstillSøknadOvergangsstønad,
+    fjernBarnFraSøknad,
   };
 });
