+++ src/skolepenger/SkolepengerContext.tsx
@@ -200,6 +200,19 @@ const [SkolepengerSøknadProvider, useSkolepengerSøknad] = createUseContext(
       }));
     };
 
+    const fjernBarnFraSøknad = (id: string) => {
+      const nyBarneListe = søknad.person.barn.filter(
+        (barn: IBarn) => barn.id !== id
+      );
+
+      settSøknad((prevSoknad: ISøknad) => {
+        return {
+          ...prevSoknad,
+          person: { ...søknad.person, barn: nyBarneListe },
+        };
+      });
+    };
+
     return {
       søknad,
       settSøknad,
@@ -212,6 +225,7 @@ const [SkolepengerSøknadProvider, useSkolepengerSøknad] = createUseContext(
       nullstillMellomlagretSkolepenger,
       nullstillSøknadSkolepenger,
       oppdaterBarnISoknaden,
+      fjernBarnFraSøknad,
     };
   }
 );
